requirements.txt
discord.py
pillow
topggpy
python-chess
pymongo[srv]
motor[srv]
